functions/sync_worker/requirements.txt
functions-framework==3.8.3
flask==3.0.3
google-cloud-firestore==2.16.1
pydantic==2.9.2
requests==2.31.0
loguru==0.7.2

